Building the basic Code for a Photovoltaics self suffiency  calculation
This change builds the first basic code for a calculation of the degree of self suffiency of a electricity consumer and a pv plant. The Code calculates the self sufficiency by simply comparing the yield of a pv plant and the consumption and choosing the smaller value which is then set as ownconsumption. The total amount of ownconsumption is divided by the total amount of power consumption to receive the total degree of self sufficiency.